ExceptionsToCopyright - allow copying connection instances
Allowing compiler generated copy ctor and assignment operator in the connection class to enable passing by value. The compiler generated copy ctor and assignment operator were added to avoid problems when we used a raw pointer for pImpl. Now we are using a `shared_ptr` so this is no longer an issue. In addition not being able to pass the `connection` by value encouraged to pass it by reference to task continuation which may result in crashes when a continuation is executed after the referenced variable went out of scope.