OnErrorResumeNext - handling a situation when receive fails during connect
Currently if an exception was thrown when waiting for the init message we would end the receive loop without faulting the start task. Given that we don't have transport connect timeout yet it would make `start()` calls hang indefinitely. After this change an exception would complete a pending `start()` call with an exception.