StutteringParser - avoiding duplicate response parsing when using hubs
`connection_impl` allowed registering a callback when a message was received. This callback used to take a string as the parameter. However for hubs the message is in json format so we had to parse the parameter. Since SignalR uses json to communicate and the response was already parsed it is more efficient to pass the json object we already had instead of serializing it and re-parsing. To fix this we add an "overload" that allows registering a callback that takes `json::value` as the parameter. The way we do it is a bit clumsy (that's why this is "overload" and overload) since there isn't a straightforward way of adding and overload taking `std::function` with different parameters. Since the `connection_impl` is an internal class this should be fine - the public API does not reflect this in any way.
The change also fixes a corner case where a hub connection would receive a persistent connection message - before we would fail when parsing a primitive value, now we can handle this case gracefully.