Fixing serializing string twice in persistent connection
Since SignalR payloads are Json we parse the incoming message as a Json object and extract the actual message. For persistent connections the api is a string based so we need to turn the message back to string. We used to always serialize it but if the value is already a string it causes double escaping. The fix is to check if the value is a string and if it is just pass it down to the receive handler, otherwise (e.g. it is a json object or a number etc.) we serialize it back to string as we did before.

Issue: #128